requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0.0
langchain
langchain-core
langchain-openai
chromadb
pypdf
python-dotenv
rank_bm25